# docs/api.yaml
openapi: 3.0.3
info:
  title: Veb Forum API
  version: 1.0.0
  description: |
    Detaljna REST API dokumentacija za Laravel aplikaciju Veb Forum.

    Za JSON zahteve koristiti:
    - Accept: application/json
    - Content-Type: application/json

    Za zaštićene rute dodatno koristiti:
    - Authorization: Bearer {token}
servers:
  - url: http://127.0.0.1:8000/api
    description: Lokalni razvojni server
...
paths:
...
  /forgot-password:
    post:
      tags: [Password]
      summary: Zahtev za reset lozinke
      description: |
        Pokreće proces promene zaboravljene lozinke za prosleđenu email adresu.

        Potrebna zaglavlja:
        - Accept: application/json
        - Content-Type: application/json
      operationId: forgotPassword
      parameters:
        - $ref: '#/components/parameters/AcceptHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: masa@example.com
      responses:
        '200':
          description: Zahtev za reset lozinke je obrađen.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /reset-password:
    post:
      tags: [Password]
      summary: Resetovanje lozinke
      description: |
        Postavlja novu lozinku na osnovu email adrese i tokena za reset.

        Potrebna zaglavlja:
        - Accept: application/json
        - Content-Type: application/json
      operationId: resetPassword
      parameters:
        - $ref: '#/components/parameters/AcceptHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Lozinka je uspešno promenjena.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: Sanctum token dobijen nakon login ili register zahteva.
  parameters:
    AcceptHeader:
      in: header
      name: Accept
      required: true
      description: Format odgovora koji klijent očekuje.
      schema:
        type: string
        default: application/json
        enum: [application/json]
...
  schemas:
...
    ResetPasswordRequest:
      type: object
      required: [email, token, password, password_confirmation]
      properties:
        email:
          type: string
          format: email
          example: masa@example.com
        token:
          type: string
          example: reset-token-example
        password:
          type: string
          format: password
          minLength: 8
          example: newpassword123
        password_confirmation:
          type: string
          format: password
          example: newpassword123
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: Operacija je uspešno izvršena.
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
...
    ValidationError:
      description: Prosleđeni podaci nisu validni.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            message: The given data was invalid.
            errors:
              title:
                - The title field is required.